Rewrite in Python
The original JavaScript implementation would attempt to read past the end of the
checklist; if the example Jenkins commands were removed, this could cause the
check to fail with a very unhelpful error message. While fixing the bug, I saw
no reason at all that this should be implemented in JavaScript. Since we have
more in-house Python expertise, it seemed like a good choice.

Signed-off-by: Zack Cerza <zack@cerza.org>